+++ packages/onionprobe/descriptor.py
@@ -177,6 +177,16 @@ class OnionprobeDescriptor:
             self.set_metric('onion_service_descriptor_latency_seconds',
                             elapsed, labels)
 
+            # Register HSDir latency
+            [ hsdir_id, hsdir_name ] = str(self.hsdirs[config['address']]).split('~')
+
+            #self.log('HSDir ID: {}, HSDir name: {}'.format(hsdir_id, hsdir_name))
+            self.set_metric('hsdir_latency_seconds',
+                            elapsed, {
+                                'name': hsdir_name,
+                                'id'  : hsdir_id,
+                                })
+
             # Debuging the outer layer
             self.log("Outer wrapper descriptor layer contents (decrypted):\n" + str(descriptor), 'debug')
 
@@ -192,7 +202,7 @@ class OnionprobeDescriptor:
             inner = descriptor.decrypt(pubkey)
 
             if descriptor.lifetime:
-                self.log("Descriptor lifetime : " + str(descriptor.lifetime))
+                self.log("Descriptor lifetime: " + str(descriptor.lifetime))
                 self.set_metric('onion_service_descriptor_lifetime',
                                 descriptor.lifetime, labels)
 
@@ -287,3 +297,10 @@ class OnionprobeDescriptor:
             'state': reason,
             },
             labels)
+
+        # Initialize the HSDirs object if needed
+        if 'hsdirs' not in dir(self):
+            self.hsdirs = {}
+
+        # Register the HSDir where the descriptor was fetched
+        self.hsdirs[event.address + '.onion'] = str(event.directory).split('$')[1]